Move ContextTypes to another file
